requirements.txt
# HTTP 클라이언트
httpx>=0.25.0
websockets>=12.0

# 데이터베이스
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Redis
redis>=5.0.0

# 스케줄러
apscheduler>=3.10.0

# LLM
anthropic>=0.40.0
# openai>=1.0.0

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0
